Refactroed util functions, adding similarity score function
